Added gpiokeyslistener for listening key evens from gpip-keys. This is needed to add support for focus (half pressed camera key) and lens cover key events. Closing/opening lens cover switches camera UI to background/foreground. For now quitting camera UI (Quit button) closes the app but in the future it will switch the process to background mode and close only the UI part.
Added local socket server to ensure that only one instance ot the
app is running at once. If new one is tried to be launched the
existing instance is popped to foreground and the new instance
quits.

Added default capture resolution setting to config file. Default
capture resolution is 1296x976.